added amino acid capabilities to fasta2tree_w.py

+++ fasta2tree_w.py
@@ -9,11 +9,14 @@ parser = argparse.ArgumentParser(description='This script takes a directory of f
 parser.add_argument('-D', dest = 'DIR', type = str, required=True,  help = 'Path to input directory of fasta files')
 parser.add_argument('-E', dest = 'ending', type = str, required=True,  help = 'Fasta files ending')
 parser.add_argument('--iqtree', dest= 'iqtree', action ='store_true', default= False, help ='use iqtree to estimate phylogeny with modeltesting (TESTNEW), default = False.')
+parser.add_argument('--aa', dest= 'aa', action ='store_true', default= False, help ='input data are amino acid sequences, automatically uses mafft for alignment and iqtree to infer phylogeny, default = False.')
 args = parser.parse_args()
 
 translatorx_path = '/Users/dbstern/Desktop/Phylogenetics_Programs/translatorx_vLocal.pl'
 raxml_cmd = 'raxmlHPC-PTHREADS-AVX -T 4'
 iqtree_cmd = 'iqtree'
+mafft_cmd = 'mafft'
+
 
 def transx(fasta_file,DIR):
 	transout = fasta_file.split('.')[0]+".transx"
@@ -62,23 +65,63 @@ def tree(alignment,DIR):
 		command = 'FastTree -gtr -nt '+DIR+alignment+' > '+DIR+cluster+'.fasttree.tre'
 		os.system(command)
 
+def tree_aa(alignment,DIR):
+	ali = AlignIO.read(DIR+alignment, "fasta")
+	if seqcount(ali) < 500:
+		cluster = alignment.split('.')[0]
+		command = iqtree_cmd+' -s '+DIR+alignment+' -m TEST -nt AUTO -pre '+DIR+cluster
+		os.system(command)
+		tree = DIR+cluster+".iqtree.tre"
+		raw_tree = DIR+cluster+'.treefile'
+		try:
+			os.rename(raw_tree,tree)
+			os.system('rm '+DIR+cluster+'.best_scheme '+DIR+cluster+".best_scheme.nex "+DIR+cluster+".bionj "+DIR+cluster+".ckp.gz "+DIR+cluster+".iqtree "+DIR+cluster+".log "+DIR+cluster+".mldist "+DIR+cluster+".model.gz")
+		except:pass 
+	else:
+		cluster = alignment.split('.')[0]
+		command = 'FastTree '+DIR+alignment+' > '+DIR+cluster+'.fasttree.tre'
+		os.system(command)
+
+def mafft_aa(fasta_file,DIR):
+	mafftout = fasta_file.split('.')[0]+".mafft"
+	command = mafft_command+" --auto "+DIR+fasta_file+" > "+DIR+mafftout
+	print "executing: " + command
+	os.system(command)
+
+
 if __name__ == "__main__":
 	DIR = args.DIR
 	ending = args.ending
 	if DIR[-1] != "/": DIR += "/"
 	for fasta_file in os.listdir(DIR):
-		if fasta_file.endswith(ending):
-			cluster = fasta_file.split('.')[0]
-			if os.path.isfile(DIR+cluster+'.transx.nt_ali.fasta'):
-				print('Detected alignment for '+cluster)
-			else:
-				transx(fasta_file,DIR)
-				os.system('rm '+DIR+'*.html '+DIR+'*nt1_ali.fasta '+DIR+'*nt2_ali.fasta '+DIR+'*nt3_ali.fasta '+DIR+'*nt12_ali.fasta '+DIR+'*.log '+DIR+'*.aaseqs '+DIR+'*.aaseqs.fasta '+DIR+'*aa_based_codon* '+DIR+'*.aa_ali.fasta')
+		if args.aa:
+			if fasta_file.endswith(ending):
+				cluster = fasta_file.split('.')[0]
+				if os.path.isfile(DIR+cluster+'.mafft'):
+					print('Detected alignment for '+cluster)
+				else:
+					maftt_aa(fasta_file,DIR)
+		else:
+			if fasta_file.endswith(ending):
+				cluster = fasta_file.split('.')[0]
+				if os.path.isfile(DIR+cluster+'.transx.nt_ali.fasta'):
+					print('Detected alignment for '+cluster)
+				else:
+					transx(fasta_file,DIR)
+					os.system('rm '+DIR+'*.html '+DIR+'*nt1_ali.fasta '+DIR+'*nt2_ali.fasta '+DIR+'*nt3_ali.fasta '+DIR+'*nt12_ali.fasta '+DIR+'*.log '+DIR+'*.aaseqs '+DIR+'*.aaseqs.fasta '+DIR+'*aa_based_codon* '+DIR+'*.aa_ali.fasta')
 	for fasta_file in os.listdir(DIR):
-		if fasta_file.endswith(ending):
-			cluster = fasta_file.split('.')[0]
-			if os.path.isfile(DIR+cluster+'.raxml.tre') or os.path.isfile(DIR+cluster+'.fasttree.tre') or os.path.isfile(DIR+cluster+'.iqtree.tre'): 
-				print('Detected tree file for '+cluster)
-			else:
+		if args.aa:
+			if fasta_file.endswith(ending):
+				cluster = fasta_file.split('.')[0]
+				if os.path.isfile(DIR+cluster+'.raxml.tre') or os.path.isfile(DIR+cluster+'.fasttree.tre') or os.path.isfile(DIR+cluster+'.iqtree.tre'): 
+					print('Detected tree file for '+cluster)
+				else:
+				tree_aa(cluster+"mafft",DIR)
+		else:
+			if fasta_file.endswith(ending):
+				cluster = fasta_file.split('.')[0]
+				if os.path.isfile(DIR+cluster+'.raxml.tre') or os.path.isfile(DIR+cluster+'.fasttree.tre') or os.path.isfile(DIR+cluster+'.iqtree.tre'): 
+					print('Detected tree file for '+cluster)
+				else:
 				partition(cluster+".transx.nt_ali.fasta",DIR)
 				tree(cluster+".transx.nt_ali.fasta",DIR)
